fullstack/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
sentence-transformers==3.2.0
scikit-learn==1.5.2
numpy==1.26.4
google-generativeai==0.8.3
